Add files via upload
The program facilitates file recovery and processing by copying files from a specified disk to an input folder, categorizing them based on file types, and renaming them with the original extensions before copying to an output folder. It does a rip in binary, which enables it to see files marked as deleted as well as normal files, and supports fat32, ntfs, and ext4 file structures. Made entirely in batch.